Export function to get PSMX loan
This exports the (internal) function to get a PSMX loan from a serdata
or a raw sample, which can be useful for specific scenarios.
